split traffic from recorder
